+++ include/ensmallen_bits/ada_grad/ada_grad_impl.hpp
@@ -23,7 +23,8 @@ inline AdaGrad::AdaGrad(const double stepSize,
                         const size_t maxIterations,
                         const double tolerance,
                         const bool shuffle,
-                        const bool resetPolicy) :
+                        const bool resetPolicy,
+                        const bool exactObjective) :
     optimizer(stepSize,
               batchSize,
               maxIterations,
@@ -31,7 +32,8 @@ inline AdaGrad::AdaGrad(const double stepSize,
               shuffle,
               AdaGradUpdate(epsilon),
               NoDecay(),
-              resetPolicy)
+              resetPolicy,
+              exactObjective)
 { /* Nothing to do. */ }
 
 } // namespace ens