+++ include/ensmallen_bits/padam/padam.hpp
@@ -64,6 +64,8 @@ class Padam
    *        function is visited in linear order.
    * @param resetPolicy If true, parameters are reset before every Optimize
    *        call; otherwise, their values are retained.
+   * @param exactObjective Calculate the exact objective (Default: estimate the
+   *        final objective obtained on the last pass over the data).
    */
   Padam(const double stepSize = 0.001,
         const size_t batchSize = 32,
@@ -74,7 +76,8 @@ class Padam
         const size_t maxIterations = 100000,
         const double tolerance = 1e-5,
         const bool shuffle = true,
-        const bool resetPolicy = true) :
+        const bool resetPolicy = true,
+        const bool exactObjective = false) :
       optimizer(stepSize,
                 batchSize,
                 maxIterations,
@@ -82,7 +85,8 @@ class Padam
                 shuffle,
                 PadamUpdate(epsilon, beta1, beta2, partial),
                 NoDecay(),
-                resetPolicy)
+                resetPolicy,
+                exactObjective)
   { /* Nothing to do here. */ }
 
   /**
@@ -145,6 +149,11 @@ class Padam
   bool Shuffle() const { return optimizer.Shuffle(); }
   //! Modify whether or not the individual functions are shuffled.
   bool& Shuffle() { return optimizer.Shuffle(); }
+
+  //! Get whether or not the actual objective is calculated.
+  bool ExactObjective() const { return optimizer.ExactObjective(); }
+  //! Modify whether or not the actual objective is calculated.
+  bool& ExactObjective() { return optimizer.ExactObjective(); }
 
   //! Get whether or not the update policy parameters
   //! are reset before Optimize call.