+++ include/ensmallen_bits/qhadam/qhadam_impl.hpp
@@ -28,7 +28,8 @@ inline QHAdam::QHAdam(
     const size_t maxIterations,
     const double tolerance,
     const bool shuffle,
-    const bool resetPolicy) :
+    const bool resetPolicy,
+    const bool exactObjective) :
     optimizer(stepSize,
               batchSize,
               maxIterations,
@@ -36,7 +37,8 @@ inline QHAdam::QHAdam(
               shuffle,
               QHAdamUpdate(epsilon, beta1, beta2, v1, v2),
               NoDecay(),
-              resetPolicy)
+              resetPolicy,
+              exactObjective)
 { /* Nothing to do. */ }
 
 } // namespace ens