+++ include/ensmallen_bits/sarah/sarah.hpp
@@ -67,6 +67,8 @@ class SARAHType
    *     function is visited in linear order.
    * @param updatePolicy Instantiated update policy used to adjust the given
    *     parameters.
+   * @param exactObjective Calculate the exact objective (Default: estimate the
+   *        final objective obtained on the last pass over the data).
    */
   SARAHType(const double stepSize = 0.01,
             const size_t batchSize = 32,
@@ -74,7 +76,8 @@ class SARAHType
             const size_t innerIterations = 0,
             const double tolerance = 1e-5,
             const bool shuffle = true,
-            const UpdatePolicyType& updatePolicy = UpdatePolicyType());
+            const UpdatePolicyType& updatePolicy = UpdatePolicyType(),
+            const bool exactObjective = false);
 
   /**
    * Optimize the given function using SARAH. The given starting point will be
@@ -119,6 +122,11 @@ class SARAHType
   //! Modify whether or not the individual functions are shuffled.
   bool& Shuffle() { return shuffle; }
 
+  //! Get whether or not the actual objective is calculated.
+  bool ExactObjective() const { return exactObjective; }
+  //! Modify whether or not the actual objective is calculated.
+  bool& ExactObjective() { return exactObjective; }
+
   //! Get the update policy.
   const UpdatePolicyType& UpdatePolicy() const { return updatePolicy; }
   //! Modify the update policy.
@@ -143,6 +151,9 @@ class SARAHType
   //! Controls whether or not the individual functions are shuffled when
   //! iterating.
   bool shuffle;
+
+  //! Controls whether or not the actual Objective value is calculated.
+  bool exactObjective;
 
   //! The update policy used to update the parameters in each iteration.
   UpdatePolicyType updatePolicy;