+++ include/ensmallen_bits/spalera_sgd/spalera_sgd_impl.hpp
@@ -30,13 +30,15 @@ SPALeRASGD<DecayPolicyType>::SPALeRASGD(const double stepSize,
                                         const double adaptRate,
                                         const bool shuffle,
                                         const DecayPolicyType& decayPolicy,
-                                        const bool resetPolicy) :
+                                        const bool resetPolicy,
+                                        const bool exactObjective) :
     stepSize(stepSize),
     batchSize(batchSize),
     maxIterations(maxIterations),
     tolerance(tolerance),
     lambda(lambda),
     shuffle(shuffle),
+    exactObjective(exactObjective),
     updatePolicy(SPALeRAStepsize(alpha, epsilon, adaptRate)),
     decayPolicy(decayPolicy),
     resetPolicy(resetPolicy)
@@ -150,12 +152,15 @@ double SPALeRASGD<DecayPolicyType>::Optimize(DecomposableFunctionType& function,
   Info << "SPALeRA SGD: maximum iterations (" << maxIterations
       << ") reached; terminating optimization." << std::endl;
 
-  // Calculate final objective.
-  overallObjective = 0;
-  for (size_t i = 0; i < numFunctions; i += batchSize)
+  // Calculate final objective if exactObjective is set to true.
+  if (exactObjective)
   {
-    const size_t effectiveBatchSize = std::min(batchSize, numFunctions - i);
-    overallObjective += f.Evaluate(iterate, i, effectiveBatchSize);
+    overallObjective = 0;
+    for (size_t i = 0; i < numFunctions; i += batchSize)
+    {
+      const size_t effectiveBatchSize = std::min(batchSize, numFunctions - i);
+      overallObjective += f.Evaluate(iterate, i, effectiveBatchSize);
+    }
   }
 
   return overallObjective;