+++ include/ensmallen_bits/wn_grad/wn_grad.hpp
@@ -58,13 +58,16 @@ class WNGrad
    *        function is visited in linear order.
    * @param resetPolicy If true, parameters are reset before every Optimize
    *        call; otherwise, their values are retained.
+   * @param exactObjective Calculate the exact objective (Default: estimate the
+   *        final objective obtained on the last pass over the data).
    */
   WNGrad(const double stepSize = 0.562,
          const size_t batchSize = 32,
          const size_t maxIterations = 100000,
          const double tolerance = 1e-5,
          const bool shuffle = true,
-         const bool resetPolicy = true);
+         const bool resetPolicy = true,
+         const bool exactObjective = false);
 
   /**
    * Optimize the given function using WNGrad. The given starting point will
@@ -106,6 +109,11 @@ class WNGrad
   bool Shuffle() const { return optimizer.Shuffle(); }
   //! Modify whether or not the individual functions are shuffled.
   bool& Shuffle() { return optimizer.Shuffle(); }
+
+  //! Get whether or not the actual objective is calculated.
+  bool ExactObjective() const { return optimizer.ExactObjective(); }
+  //! Modify whether or not the actual objective is calculated.
+  bool& ExactObjective() { return optimizer.ExactObjective(); }
 
   //! Get whether or not the update policy parameters
   //! are reset before Optimize call.